decoupling and implementing single responsibility pattern
