+++ lib/tic_tac_toe_game/greeting.rb
@@ -1,0 +1,7 @@
+module TicTacToeGame
+    class Greeting
+      def self.welcome
+          'Welcome to Tic Tac Toe Game'
+      end
+    end
+end
