+++ lib/tic_tac_toe_game/move.rb
@@ -1,0 +1,16 @@
+module TicTacToeGame
+    class Move      
+        def initialize(board, turn, movelist)
+            @board = board
+            @turn = turn
+            @movelist = movelist
+        end
+
+        def move idx
+            @board[idx] = @turn
+            @turn = other_turn
+            @movelist << idx
+            self
+        end
+    end
+end
