+++ lib/tic_tac_toe_game/play.rb
@@ -1,0 +1,11 @@
+module TicTacToeGame
+    class Play      
+        def initialize(board)
+            @board = board
+        end
+
+        def possible_moves
+            @board.map.with_index { |piece, idx| piece == "-" ? idx : nil }.compact
+        end
+    end
+end
