+++ spec/tic_tac_toe_game/greeting_spec.rb
@@ -1,0 +1,5 @@
+require_relative '../../lib/tic_tac_toe_game/greeting.rb'
+
+RSpec.describe TicTacToeGame::Greeting do
+    it { expect(TicTacToeGame::Greeting.welcome).to eql 'Welcome to Tic Tac Toe Game' }
+end
